user page setting

+++ app/controllers/cart_items/orders_controller.rb
@@ -3,6 +3,7 @@ class CartItems::OrdersController < ApplicationController
   def new
     @order = Order.new
     @order_item = OrderItem.new
+    @cart_items = current_user.cart_items
   end
 
   def create
@@ -11,20 +12,23 @@ class CartItems::OrdersController < ApplicationController
     @order.user_id = current_user.id
     @order.shop_id = @shop.id
     @order.numbering_reserve_number(@order.shop.target_date_order_count(@order.takeaway_datetime))
-    @order.save!
-    @cart_items = current_user.cart_items
-    @cart_items.each do |cart_item|
-      OrderItem.create!(
-        order_id: @order.id,
-        menu_id: cart_item.menu.id,
-        menu_name: cart_item.menu.name,
-        menu_price: cart_item.menu.price,
-        menu_amount: cart_item.amount
-      )
-      cart_item.destroy
+    if @order.save
+      @cart_items = current_user.cart_items
+      @cart_items.each do |cart_item|
+        OrderItem.create(
+          order_id: @order.id,
+          menu_id: cart_item.menu.id,
+          menu_name: cart_item.menu.name,
+          menu_price: cart_item.menu.price,
+          menu_amount: cart_item.amount
+        )
+        cart_item.destroy
+      end
+      redirect_to orders_fix_path
+    else
+      flas.now[:danger] = "注文できませんでした。受け取り時間を確認してください。"
+      render :new
     end
-    flash[:success] = "注文しました"
-    redirect_to menus_path
   end
 
   private
